+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+* (Change) Adds 'v$' as an 'id-length' exception.
+
 ## 16.0.1
 * (Breaking) `@typescript-eslint/method-signature-style` enforces the method style in interfaces.
 * (Breaking) Enforces `max-lines` by default.